// src/app/features/patient/models/patient.ts
export interface Patient {

  id: number;

  firstName: string;

  lastName: string;

  dateOfBirth: string;

  gender: string;

  bloodGroup: string;

  phoneNumber: string;

  email: string;

  address: string;

  emergencyContactName: string;

  emergencyContactPhone: string;

  insuranceNumber: string;

  isActive: boolean;

}