// src/app/features/patient/models/update-patient.ts
export interface UpdatePatient {

  id: number;

  firstName: string;

  lastName: string;

  dateOfBirth: string;

  gender: number;

  bloodGroup: number;

  phoneNumber: string;

  email: string;

  address: string;

  emergencyContactName: string;

  emergencyContactPhone: string;

  insuranceNumber: string;

}