// src/app/features/patient/models/create-patient.ts
export interface CreatePatient {

  firstName: string;

  lastName: string;

  dateOfBirth: string;

  gender: number;

  bloodGroup: number;

  phoneNumber: string;

  email: string;

  address: string;

  emergencyContactName: string;

  emergencyContactPhone: string;

  insuranceNumber: string;

}